pyproject.toml
[tool.poetry]
name = "appium-day-1"
version = "0.1.0"
description = ""
authors = ["Sanjar <sabrorxojaev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.16.0"
pytest = "^7.4.3"
appium-python-client = "^3.1.1"
pillow = "^10.2.0"
pytest-check = "^2.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
